requirements.txt
Flask
gunicorn
flake8
pyyaml
